fix: replace getAlltag
