+++ src/components/elements/taskmanager.tsx
@@ -2,7 +2,7 @@
 import { useEffect, useState } from 'react';
 import { getCookie } from 'cookies-next';
 import BASE_URL, { type TaskManageMentOverviewProp } from '@/lib/shared';
-import type { TaskProps, TagProps } from '@/app/types/types';
+import type { TaskProps } from '@/app/types/types';
 import { Task, ExportDialog, Filter, Sort, CreateTask } from './taskManagement';
 import { parseJsonValues, statusSections } from '@/lib/taskUtils';
 
@@ -13,7 +13,6 @@ export const TaskManager = ({ project_id }: TaskManageMentOverviewProp) => {
   const [tasks, setTasks] = useState<TaskProps[]>([]);
   const [showTasks, setShowTasks] = useState<TaskProps[]>([]);
   const [projectName, setProjectName] = useState<string>('');
-  const [allTags, setAllTags] = useState<TagProps[]>([]);
 
   useEffect(() => {
     //get all data of project from db
@@ -36,25 +35,6 @@ export const TaskManager = ({ project_id }: TaskManageMentOverviewProp) => {
         console.error(error);
       }
     };
-    //get all tags of tasks from db
-    const fetchTagData = async () => {
-      const url = `${BASE_URL}/v2/tags/`;
-      const options = {
-        method: 'GET',
-        headers: {
-          Authorization: auth,
-        },
-      };
-
-      try {
-        const response = await fetch(url, options);
-        const data = (await response.json()) as TagProps[];
-        setAllTags(data);
-      } catch (error) {
-        console.error(error);
-      }
-    };
-    fetchTagData();
     fetchData();
   }, [project_id]);
 
@@ -62,7 +42,7 @@ export const TaskManager = ({ project_id }: TaskManageMentOverviewProp) => {
     return (
       <div className="flex items-center justify-between w-full mb-3">
         <div className="flex items-center gap-4">
-          <Filter tasks={tasks} allTags={allTags} setShowTasks={setShowTasks} />
+          <Filter tasks={tasks} setShowTasks={setShowTasks} />
           <ExportDialog tasks={tasks} />
         </div>
         <div className="flex items-center gap-4">
